Refactor operators, use new BaseOperator class
