+++ docs/dictionary/wordlist.txt
@@ -62,6 +62,7 @@ bambara
 barcode
 basecall
 basecoord
+baseoperator
 baseurl
 bashkir
 bcf
@@ -462,6 +463,7 @@ objectdetails
 objectid
 objectinsidesubmission
 objectinsidesubmissionwithtags
+objectoperator
 objectstatus
 objectsubmissiontypes
 objecttags
