+++ metadata_backend/api/operators/submission.py
@@ -2,31 +2,19 @@
 import re
 from datetime import datetime
 from typing import Dict, List, Optional, Tuple, Union
-from uuid import uuid4
 
 from aiohttp import web
-from motor.motor_asyncio import AsyncIOMotorClient
 from pymongo.errors import ConnectionFailure, OperationFailure
 
-from ...conf.conf import mongo_database
-from ...database.db_service import DBService
 from ...helpers.logger import LOG
+from .base import BaseOperator
 
 
-class SubmissionOperator:
-    """Operator class for handling database operations of submissions.
+class SubmissionOperator(BaseOperator):
+    """ObjectOperator class for handling database operations of submissions.
 
     Operations are implemented with JSON format.
     """
-
-    def __init__(self, db_client: AsyncIOMotorClient) -> None:
-        """Init db_service.
-
-        :param db_client: Motor client used for database connections. Should be
-        running on same loop with aiohttp, so needs to be passed from aiohttp
-        Application.
-        """
-        self.db_service = DBService(mongo_database, db_client)
 
     async def get_submission_field(self, submission_id: str, field: str) -> Union[str, list, dict]:
         """Get a field from the submission.
@@ -171,7 +159,7 @@ class SubmissionOperator:
         :raises: HTTPBadRequest if error occurs during the process of insert
         :returns: Submission id for the submission inserted to database
         """
-        submission_id = self._generate_submission_id()
+        submission_id = self._generate_accession_id()
         _now = int(datetime.now().timestamp())
         data["submissionId"] = submission_id
         data["text_name"] = " ".join(re.split("[\\W_]", data["name"]))
@@ -353,12 +341,3 @@ class SubmissionOperator:
             reason = f"Submission with ID: '{submission_id}' is already published and cannot be modified or deleted."
             LOG.error(reason)
             raise web.HTTPMethodNotAllowed(method=method, allowed_methods=["GET", "HEAD"], reason=reason)
-
-    def _generate_submission_id(self) -> str:
-        """Generate random submission id.
-
-        :returns: str with submission id
-        """
-        sequence = uuid4().hex
-        LOG.debug("Generated submission ID.")
-        return sequence
