+++ metadata_backend/api/operators/user.py
@@ -1,34 +1,22 @@
 """User operator class."""
 from typing import Dict, List, Tuple, Union
-from uuid import uuid4
 
 import aiohttp_session
 from aiohttp import web
-from motor.motor_asyncio import AsyncIOMotorClient
 from pymongo.errors import ConnectionFailure, OperationFailure
 
-from ...conf.conf import mongo_database
-from ...database.db_service import DBService
 from ...helpers.logger import LOG
 from ...helpers.validator import JSONValidator
-from .object import Operator
+from .base import BaseOperator
+from .object import ObjectOperator
 from .submission import SubmissionOperator
 
 
-class UserOperator:
-    """Operator class for handling database operations of users.
+class UserOperator(BaseOperator):
+    """ObjectOperator class for handling database operations of users.
 
     Operations are implemented with JSON format.
     """
-
-    def __init__(self, db_client: AsyncIOMotorClient) -> None:
-        """Init db_service.
-
-        :param db_client: Motor client used for database connections. Should be
-        running on same loop with aiohttp, so needs to be passed from aiohttp
-        Application.
-        """
-        self.db_service = DBService(mongo_database, db_client)
 
     async def check_user_has_doc(
         self, req: web.Request, collection: str, user_id: str, accession_id: str
@@ -55,7 +43,7 @@ class UserOperator:
 
         project_id = ""
         if collection.startswith("template"):
-            object_operator = Operator(db_client)
+            object_operator = ObjectOperator(db_client)
             project_id = await object_operator.get_object_project(collection, accession_id)
         elif collection == "submission":
             submission_operator = SubmissionOperator(db_client)
@@ -110,7 +98,7 @@ class UserOperator:
                 return existing_user_id
 
             user_data["projects"] = data["projects"]
-            user_data["userId"] = user_id = self._generate_user_id()
+            user_data["userId"] = user_id = self._generate_accession_id()
             user_data["name"] = data["real_name"]
             user_data["externalId"] = data["user_id"]
             JSONValidator(user_data, "users")
@@ -200,12 +188,3 @@ class UserOperator:
             reason = f"User with ID: '{user_id}' was not found."
             LOG.error(reason)
             raise web.HTTPNotFound(reason=reason)
-
-    def _generate_user_id(self) -> str:
-        """Generate random user id.
-
-        :returns: str with user id
-        """
-        sequence = uuid4().hex
-        LOG.debug("Generated user ID.")
-        return sequence
